% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doppelganger.R
\name{inspect_corrs}
\alias{inspect_corrs}
\title{Extract Correlations}
\usage{
inspect_corrs(dg_object, variable)
}
\arguments{
\item{dg_object}{An object of class \code{doppelganger}.}

\item{variable}{The variable of interest.}
}
\description{
Given one or more variables, extracts the correlations with other
variables in the source data set.
}
