% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{measures}
\alias{measures}
\title{Simulated Measures}
\format{
A tibble containing 100 and 7 variables.
}
\usage{
measures
}
\description{
Simulated data matrix to show how the package \code{doppelganger} works.
}
\keyword{data}
