% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doppelganger.R
\name{doppelganger}
\alias{doppelganger}
\title{Find Doppelgangers}
\usage{
doppelganger(
  data,
  variables = NULL,
  priority = c("centrality", "raw_order"),
  threshold = 0.9
)
}
\arguments{
\item{data}{Data table source.}

\item{variables}{Variables included in \code{data} to consider.}

\item{priority}{Ranking method to prioritize variables. When "Centrality" is
selected, variables with higher mean absolute correlation will have higher
priority. "Raw order" prioritize variables using their position within the
vector \code{variables}.}

\item{threshold}{Correlation cut-off (absolute value) to identify doppelgangers.
Variables with absolute correlation values equal or greater to this value will be
considered aliases of each other.}
}
\description{
Analyze correlations to select a subset of non-redundant variables.
}
\details{
The variable priority is established by calculating a centrality index
for each variable, and sorting then accordingly.
}
\examples{
\dontrun{
data(measures)
dg <- doppelganger(measures, threshold=0.7)
qgraph::qgraph(cor(measures), minimum=0.65, edge.labels=TRUE, layout="spring")
}
}
